+++ src/main/java/me/towdium/jecalculation/data/label/ILabel.java
@@ -12,7 +12,6 @@ import me.towdium.jecalculation.gui.guis.pickers.PickerSimple;
 import me.towdium.jecalculation.utils.Utilities;
 import me.towdium.jecalculation.utils.Utilities.ReversedIterator;
 import me.towdium.jecalculation.utils.wrappers.Pair;
-import mezz.jei.api.gui.IRecipeLayout;
 import net.minecraft.enchantment.EnchantmentData;
 import net.minecraft.fluid.Fluids;
 import net.minecraft.item.ItemStack;
@@ -246,7 +245,7 @@ public interface ILabel {
             else if (o instanceof ItemStack) return new LItemStack((ItemStack) o);
             else if (o instanceof FluidStack) return new LFluidStack((FluidStack) o);
             else if (o instanceof EnchantmentData) return new LItemStack(getEnchantedItemStack((EnchantmentData) o));
-            else return new LPlaceholder(o.toString(), 1);
+            else return LPlaceholder.Converter.from(o);
         }
 
         public void register(ConverterFunction handler, Priority priority) {
@@ -254,13 +253,13 @@ public interface ILabel {
         }
 
         // get most possible guess from labels
-        public ILabel first(List<ILabel> labels, @Nullable IRecipeLayout context) {
+        public ILabel first(List<ILabel> labels, @Nullable Class<?> context) {
             List<ILabel> guess = guess(labels, context).one;
             return guess.isEmpty() ? labels.get(0) : guess.get(0);
         }
 
         // to test if the labels can be converted to other labels (like oreDict)
-        public Pair<List<ILabel>, List<ILabel>> guess(List<ILabel> labels, @Nullable IRecipeLayout context) {
+        public Pair<List<ILabel>, List<ILabel>> guess(List<ILabel> labels, @Nullable Class<?> context) {
             List<ILabel> suggest = new ReversedIterator<>(handlers.get(Priority.SUGGEST)).stream()
                     .flatMap(h -> h.convert(labels, context).stream())
                     .collect(Collectors.toList());
@@ -272,7 +271,7 @@ public interface ILabel {
 
         @FunctionalInterface
         public interface ConverterFunction {
-            List<ILabel> convert(List<ILabel> a, @Nullable IRecipeLayout context);
+            List<ILabel> convert(List<ILabel> a, @Nullable Class<?> context);
         }
     }
 
@@ -481,8 +480,18 @@ public interface ILabel {
 
         @Override
         public void getToolTip(List<String> existing, boolean detailed) {
-            if (detailed) existing.add(FORMAT_GREY +
-                    Utilities.I18n.get("label.common.amount", getAmountString(false)));
+            if (detailed) {
+                long amount = percent ? (getAmount() + 99) / 100 : getAmount();
+                long stack = amount / 64;
+                long remainder = amount % 64;
+                boolean showStack = stack != 0;
+                boolean showReminder = stack < 1000 && remainder != 0;
+                String tip = "";
+                if (showStack) tip += Utilities.cutNumber(stack, 5) + "x64";
+                if (showStack && showReminder) tip += "+";
+                if (showReminder) tip += Long.toString(remainder);
+                existing.add(FORMAT_GREY + Utilities.I18n.get("label.common.amount", tip));
+            }
         }
 
         @Override
@@ -548,9 +557,5 @@ public interface ILabel {
         public boolean matches(Object l) {
             return l instanceof Impl && percent == ((Impl) l).percent;
         }
-
-        //abstract protected void drawLabel(JecaGui gui);
-
-
     }
 }