+++ src/main/java/me/towdium/jecalculation/data/label/labels/LPlaceholder.java
@@ -6,6 +6,7 @@ import me.towdium.jecalculation.data.label.ILabel;
 import me.towdium.jecalculation.gui.JecaGui;
 import me.towdium.jecalculation.gui.Resource;
 import me.towdium.jecalculation.utils.Utilities;
+import me.towdium.jecalculation.utils.wrappers.Pair;
 import net.minecraft.nbt.CompoundNBT;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
@@ -15,7 +16,12 @@ import net.minecraftforge.fml.common.Mod;
 
 import javax.annotation.Nullable;
 import javax.annotation.ParametersAreNonnullByDefault;
+import java.util.Arrays;
 import java.util.List;
+import java.util.Objects;
+import java.util.function.Function;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
 /**
@@ -126,5 +132,32 @@ public class LPlaceholder extends ILabel.Impl {
             LPlaceholder lpB = (LPlaceholder) b;
             return lpA.name.equals(lpB.name);
         } else return false;
+    }
+
+    public static class Converter {
+        @SuppressWarnings("ArraysAsListWithZeroOrOneArgument")
+        static List<Pair<Pattern, Function<Matcher, LPlaceholder>>> converters = Arrays.asList(
+                new Pair<>(Pattern.compile("\\[\\[Gas: mekanism:(.+)], (\\d+)]"),
+                        i -> new LPlaceholder("Gas - " + capitalize(i.group(1)), Integer.parseInt(i.group(2))))
+        );
+
+        public static String capitalize(String s) {
+            String[] arr = s.replace('_', ' ').split(" ");
+            StringBuilder sb = new StringBuilder();
+
+            for (String value : arr) {
+                sb.append(Character.toUpperCase(value.charAt(0)));
+                sb.append(value.substring(1)).append(" ");
+            }
+            return sb.toString().trim();
+        }
+
+        public static LPlaceholder from(Object o) {
+            String s = o.toString();
+            return converters.stream().map(i -> {
+                Matcher m = i.one.matcher(s);
+                return m.matches() ? i.two.apply(m) : null;
+            }).filter(Objects::nonNull).findFirst().orElseGet(() -> new LPlaceholder(s, 1));
+        }
     }
 }