+++ src/main/java/me/towdium/jecalculation/data/structure/Recipe.java
@@ -140,6 +140,10 @@ public class Recipe {
     }
 
     public enum IO {
-        INPUT, OUTPUT, CATALYST
+        INPUT, OUTPUT, CATALYST;
+
+        public static IO isInput(boolean b) {
+            return b ? INPUT : OUTPUT;
+        }
     }
 }